# backend/turismo-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Turismo Necochea",
        "description": "Documentación de la API para la aplicación de Turismo de Necochea",
        "contact": {
            "email": "tudai@alumnos.exa.unicen.edu.ar"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtener todos los eventos",
                "operationId": "8df3c75a189cf773a3eb105ea46e145a",
                "responses": {
                    "200": {
                        "description": "Lista de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Evento"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear un nuevo evento",
                "operationId": "4594ce682158c01a68019d32f26180e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/eventos/{id}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtener un evento por ID",
                "operationId": "09ec864fa45a38fe55c3078da04e8a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualizar un evento por ID",
                "operationId": "032b9536863e8e3b817a73e44d058c1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Eliminar un evento por ID",
                "operationId": "a904dc7b49560c21fe21f57e1df8813a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento eliminado correctamente"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            }
        },
        "/api/eventos/destacados": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtener eventos destacados",
                "operationId": "11caa16cc782773011985daa91df9747",
                "responses": {
                    "200": {
                        "description": "Lista de eventos destacados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Evento"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/gastronomicos/{gastronomicoId}/menus": {
            "get": {
                "tags": [
                    "Gastronomicos"
                ],
                "summary": "Obtener los menús asociados a un establecimiento gastronómico",
                "operationId": "7b1b1346469d9d69f02af3362b71b669",
                "parameters": [
                    {
                        "name": "gastronomicoId",
                        "in": "path",
                        "description": "ID del establecimiento gastronómico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de menús asociados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Menu"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gastronómico no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Gastronomicos"
                ],
                "summary": "Asociar un menú a un establecimiento gastronómico",
                "operationId": "331c0c7ffbae562d373920d64c4f21d8",
                "parameters": [
                    {
                        "name": "gastronomicoId",
                        "in": "path",
                        "description": "ID del establecimiento gastronómico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "menu_id"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menú asociado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gastronomico"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gastronómico no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/gastronomicos/{gastronomicoId}/menus/{menuId}": {
            "delete": {
                "tags": [
                    "Gastronomicos"
                ],
                "summary": "Desasociar un menú de un establecimiento gastronómico",
                "operationId": "436082d1c1694101e562a3caf877bca6",
                "parameters": [
                    {
                        "name": "gastronomicoId",
                        "in": "path",
                        "description": "ID del establecimiento gastronómico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "menuId",
                        "in": "path",
                        "description": "ID del menú a desasociar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menú eliminado del gastronómico"
                    },
                    "404": {
                        "description": "Gastronómico no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Evento": {
                "title": "Evento",
                "description": "Modelo de un evento",
                "properties": {
                    "id": {
                        "description": "ID único del evento",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Nombre del evento",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del evento",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del evento",
                        "type": "string",
                        "nullable": true
                    },
                    "fecha": {
                        "description": "Fecha del evento (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date"
                    },
                    "lugar": {
                        "description": "Lugar donde se realiza el evento",
                        "type": "string"
                    },
                    "imagen": {
                        "description": "URL de la imagen del evento",
                        "type": "string",
                        "nullable": true
                    },
                    "destacado": {
                        "description": "Indica si el evento es destacado",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EventoRequest": {
                "title": "EventoRequest",
                "description": "Esquema para la creación de un evento",
                "required": [
                    "nombre",
                    "direccion",
                    "fecha",
                    "lugar"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Concierto de Verano"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Parque Miguel Lillo"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Música en vivo al aire libre",
                        "nullable": true
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "lugar": {
                        "type": "string",
                        "example": "Anfiteatro"
                    },
                    "imagen": {
                        "type": "string",
                        "example": "http://imagen.com/concierto.jpg",
                        "nullable": true
                    },
                    "destacado": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "EventoUpdateRequest": {
                "title": "EventoUpdateRequest",
                "description": "Esquema para la actualización de un evento",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Festival de Jazz",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Centro Cultural",
                        "nullable": true
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Noche de jazz y blues",
                        "nullable": true
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15",
                        "nullable": true
                    },
                    "lugar": {
                        "type": "string",
                        "example": "Teatro Municipal",
                        "nullable": true
                    },
                    "imagen": {
                        "type": "string",
                        "example": "http://imagen.com/jazz.jpg",
                        "nullable": true
                    },
                    "destacado": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Gastronomico": {
                "title": "Gastronomico",
                "description": "Modelo de un establecimiento gastronómico",
                "properties": {
                    "id": {
                        "description": "ID único del establecimiento",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Nombre del establecimiento",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del establecimiento",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Número de teléfono",
                        "type": "string",
                        "nullable": true
                    },
                    "redesSociales": {
                        "description": "URL de redes sociales",
                        "type": "string",
                        "nullable": true
                    },
                    "horario": {
                        "description": "Horario de atención",
                        "type": "string",
                        "nullable": true
                    },
                    "tiendaOnline": {
                        "description": "URL de la tienda online",
                        "type": "string",
                        "nullable": true
                    },
                    "extras": {
                        "description": "Información extra",
                        "type": "string",
                        "nullable": true
                    },
                    "imagen": {
                        "description": "URL de la imagen principal",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    },
                    "tipo": {
                        "description": "Tipos de gastronomía asociados (solo nombres)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "menu": {
                        "description": "Menús especiales asociados (solo nombres)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "menus": {
                        "description": "Lista completa de objetos de menú asociados",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Menu"
                        }
                    },
                    "tipos": {
                        "description": "Lista completa de objetos de tipo gastronómico asociados",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TipoGastronomico"
                        }
                    }
                },
                "type": "object"
            },
...
            "Menu": {
                "title": "Menu",
                "description": "Modelo de un tipo de menú",
                "properties": {
                    "id": {
                        "description": "ID único del menú",
                        "type": "integer",
                        "format": "int64"
                    },
                    "tipo": {
                        "description": "Tipo de menú (ej. 'Desayuno', 'Almuerzo')",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "TipoGastronomico": {
                "title": "TipoGastronomico",
                "description": "Modelo de un tipo de gastronomía",
                "properties": {
                    "id": {
                        "description": "ID único del tipo de gastronomía",
                        "type": "integer",
                        "format": "int64"
                    },
                    "tipo": {
                        "description": "Nombre del tipo de gastronomía (ej. 'Restaurante', 'Bar')",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}